feat: add Badge
